+++ CHANGELOG.md
@@ -15,6 +15,8 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 
 - [#33](https://github.com/Synesso/scala-stellar-sdk/issues/33) Support for pathfinding endpoint `/paths`.
 - [#35](https://github.com/Synesso/scala-stellar-sdk/issues/35) Support for trade aggregations endpoint `/trade_aggregations`.
+- [#53](https://github.com/Synesso/scala-stellar-sdk/issues/53) On-the-fly decoding of transaction meta-info about the
+    entries affecting the ledger.
 
 ### Changed
 - [#92](https://github.com/Synesso/scala-stellar-sdk/issues/92) Horizon Release v0.18.0 compatibility. Added fields